+++ experimental/packages/otlp-exporter-base/src/configuration/create-legacy-browser-delegate.ts
@@ -16,6 +16,7 @@
 import { ISerializer } from '@opentelemetry/otlp-transformer';
 import {
   createOtlpFetchExportDelegate,
+  createOtlpFetchLaterExportDelegate,
   createOtlpSendBeaconExportDelegate,
 } from '../otlp-browser-http-export-delegate';
 import { convertLegacyBrowserHttpOptions } from './convert-legacy-browser-http-options';
@@ -49,7 +50,13 @@ export function createLegacyOtlpBrowserExportDelegate<Internal, Response>(
 export function inferExportDelegateToUse(
   configHeaders: OTLPExporterConfigBase['headers']
 ) {
-  if (!configHeaders && typeof navigator.sendBeacon === 'function') {
+  if (
+    'fetchLater' in globalThis &&
+    /* eslint-disable  @typescript-eslint/no-explicit-any */
+    typeof (globalThis as any).fetchLater !== 'undefined'
+  ) {
+    return createOtlpFetchLaterExportDelegate;
+  } else if (!configHeaders && typeof navigator.sendBeacon === 'function') {
     return createOtlpSendBeaconExportDelegate;
   }