+++ experimental/packages/otlp-exporter-base/src/otlp-browser-http-export-delegate.ts
@@ -20,6 +20,7 @@ import { createRetryingTransport } from './retrying-transport';
 import { createSendBeaconTransport } from './transport/send-beacon-transport';
 import { createOtlpNetworkExportDelegate } from './otlp-network-export-delegate';
 import { createFetchTransport } from './transport/fetch-transport';
+import { createFetchLaterTransport } from './transport/fetch-later-transport';
 
 export function createOtlpFetchExportDelegate<Internal, Response>(
   options: OtlpHttpConfiguration,
@@ -46,6 +47,19 @@ export function createOtlpSendBeaconExportDelegate<Internal, Response>(
         url: options.url,
         headers: options.headers,
       }),
+    })
+  );
+}
+
+export function createOtlpFetchLaterExportDelegate<Internal, Response>(
+  options: OtlpHttpConfiguration,
+  serializer: ISerializer<Internal, Response>
+): IOtlpExportDelegate<Internal> {
+  return createOtlpNetworkExportDelegate(
+    options,
+    serializer,
+    createRetryingTransport({
+      transport: createFetchLaterTransport(options),
     })
   );
 }